+++ src/lex.rs
@@ -466,6 +466,7 @@ impl Lexer {
                 '∷' => self.end(Primitive::Both, start),
                 '·' => self.end(Primitive::Identity, start),
                 '⍛' => self.end(Primitive::Fill, start),
+                '⌂' => self.end(Primitive::Rise, start),
 
                 '(' => self.end(OpenParen, start),
                 ')' => self.end(CloseParen, start),